@@ -1,6 +1,6 @@
 {
   "name": "squidlet-lib",
-  "version": "0.1.10",
+  "version": "0.2.0",
   "description": "Common lib for squidlet projects",
   "author": "Ivan Kozyrin <ipkozyrin@gmail.com>",
   "repository": "git@github.com:bozonx/squidlet-lib.git",
@@ -11,8 +11,7 @@
   "exports": {
     ".": {
       "types": "./distr/types/index.d.ts",
-      "import": "./distr/lib/index.js",
-      "require": "./distr/lib/index.cjs"
+      "import": "./distr/lib/index.js"
     },
     "./package.json": "./package.json"
   },
@@ -26,16 +25,14 @@
     "exports": {
       ".": {
         "types": "./distr/types/index.d.ts",
-        "import": "./distr/lib/index.js",
-        "require": "./distr/lib/index.cjs"
+        "import": "./distr/lib/index.js"
       },
       "./package.json": "./package.json"
     }
   },
   "license": "MIT",
   "scripts": {
-    "build": "tsc && npm run build:cjs",
-    "build:cjs": "tsc --project tsconfig.cjs.json",
+    "build": "tsc",
     "test": "jest",
     "test:watch": "jest --watch",
     "coverage": "jest --coverage",
@@ -43,7 +40,7 @@
     "prepublishOnly": "npm run build"
   },
   "engines": {
-    "node": ">=18.0.0"
+    "node": ">=18.17.0"
   },
   "devDependencies": {
     "@types/jest": "^29.5.12",
